be/requirements.txt
Django==5.2
Pillow==11.2.1
djangorestframework==3.16
djangorestframework-simplejwt==5.3.0
markdown==3.8
django-filter==25.1
drf-yasg==1.21.7
